ContractDeploymentData test
