+++ yarn-project/circuits.js/src/structs/index.ts
@@ -4,6 +4,7 @@ export * from './call_context.js';
 export * from './call_request.js';
 export * from './call_stack_item.js';
 export * from './complete_address.js';
+export * from './contract_deployment_data.js';
 export * from './function_data.js';
 export * from './function_leaf_preimage.js';
 export * from './global_variables.js';
